feat(chat): 完成 Chat 对话引擎后端及 LLM 序列化重构
- 非流式/SSE 流式对话、单轮工具调用、RAG 增强
- ChatRequest 新增 ragMode 参数控制 RAG 策略
- ChatServiceImpl/LLM Client 用实体类替代 Map/ObjectNode 组装
- EmbeddingServiceImpl 用实体类替代 ObjectNode
- LlmHttpClient 新增 debug 日志打印 URL/headers/body
- 新增 15+ DTO 类（OpenAiChatResponse/ClaudeChatResponse/SseMetaEvent 等）
- 前端移除模型管理页面
- 补充 rules-snippets 代码示例备份

+++ ciff-provider/src/main/java/com/ciff/provider/dto/ModelDefaultParam.java
@@ -2,6 +2,8 @@ package com.ciff.provider.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * Model default parameters.
  * Stored as JSON in t_model.default_params column.
@@ -13,13 +15,13 @@ public class ModelDefaultParam {
      * Controls randomness. Lower values = more deterministic.
      * Range: 0.0 ~ 2.0, default varies by provider.
      */
-    private Double temperature;
+    private BigDecimal temperature;
 
     /**
      * Nucleus sampling threshold.
      * Range: 0.0 ~ 1.0.
      */
-    private Double topP;
+    private BigDecimal topP;
 
     /**
      * Maximum tokens for the completion.
@@ -31,11 +33,11 @@ public class ModelDefaultParam {
      * Frequency penalty. Positive values reduce repetition.
      * Range: -2.0 ~ 2.0.
      */
-    private Double frequencyPenalty;
+    private BigDecimal frequencyPenalty;
 
     /**
-     * Presence penalty. Positive values encourage new topics.
+     * presence penalty. positive values encourage new topics.
      * Range: -2.0 ~ 2.0.
      */
-    private Double presencePenalty;
+    private BigDecimal presencePenalty;
 }
